refactors hero page card style
