+++ features/hero-page-card/component/hero-page-card.tsx
@@ -5,20 +5,18 @@ import * as H from "./hero-page-card.style";
 type HeroPageCardProps = {
   src: string;
   alt: string;
+  title: string;
+  info: string;
 };
-export function HeroPageCard({ src, alt }: HeroPageCardProps) {
+export function HeroPageCard({ title, info, ...cardProps }: HeroPageCardProps) {
   return (
     <H.Card>
       <div>
-        <H.Image src={src} alt={alt} />
+        <H.Image {...cardProps} />
       </div>
       <H.TextandButtonWrapper>
-        <h2>Captain Marvel Vol. 1: Rise of Alpha Flight</h2>
-        <p>
-          Oh Captain, my Captain! Carol Danvers soars higher than ever with her
-          greatest mission yet — leading Earth’s first line of defense, the
-          all-new Alpha Flight space program!
-        </p>
+        <h2>{title}</h2>
+        <p>{info}</p>
       </H.TextandButtonWrapper>
       <H.ButtonWrapper>
         <PageButton href="/" label="Find Out More!" />
