+++ features/hero-page-card/index.ts
@@ -1,1 +1,3 @@
-export { HeroPageCard } from "./hero-page-card";
+export { heroPageCards } from "./api/hero-page-card-data";
+export { HeroPageCard } from "./component/hero-page-card";
+export { HeroPageCardContainer } from "./component/hero-page-card.style";
