+++ pages/index.tsx
@@ -1,20 +1,5 @@
 import Head from "next/head";
-import { HeroBanner, HeroPageCard, PageContainer } from "@features";
-import styled from "@emotion/styled";
-
-const heroPageCards = [
-  { name: "cm-card-comic1", src: "/icons/cm-card-comic1.jpg" },
-  { name: "cm-card-comic2", src: "/icons/cm-card-comic2.jpg" },
-  { name: "cm-card-comic3", src: "/icons/cm-card-comic3.jpg" },
-  { name: "cm-card-comic4", src: "/icons/cm-card-comic4.jpg" }
-];
-const HeroPageCardContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: space-around;
-  align-items: center;
-`;
+import * as H from "@features";
 
 export default function Home() {
   return (
@@ -24,14 +9,20 @@ export default function Home() {
         <meta name="description" content="Generated by create next app" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </Head>
-      <PageContainer>
-        <HeroBanner />
-        <HeroPageCardContainer>
-          {heroPageCards.map((card, index) => (
-            <HeroPageCard key={index} src={card.src} alt={card.name} />
+      <H.PageContainer>
+        <H.HeroBanner />
+        <H.HeroPageCardContainer>
+          {H.heroPageCards.map((card, index) => (
+            <H.HeroPageCard
+              key={index}
+              src={card.src}
+              alt={card.name}
+              title={card.title}
+              info={card.info}
+            />
           ))}
-        </HeroPageCardContainer>
-      </PageContainer>
+        </H.HeroPageCardContainer>
+      </H.PageContainer>
     </>
   );
 }
